requirements.txt
streamlit
numpy
pillow
scipy
matplotlib
sounddevice  # optional if you use playback locally
